requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
gtfs-realtime-bindings==0.0.7
idna==3.3
protobuf==4.21.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
six==1.16.0
urllib3==1.26.9
